% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bootdata.R
\name{create_bootdata}
\alias{create_bootdata}
\title{Create list of bootstrapped datasets using a cluster variable}
\usage{
create_bootdata(df, cluster_var, nboot, seed = NULL)
}
\arguments{
\item{df}{data.frame in "long" format.}

\item{cluster_var}{character; matches one column name in \code{df}.}

\item{nboot}{numeric; specifies how many datasets to create. Coerced to
integer.}

\item{seed}{numeric; if specified, sets seed for reproducibility.}
}
\value{
A list of length \code{nboot}, where each element is a
  \code{data.frame} including all records from each sampled value of
  \code{df$cluster_var}.
}
\description{
Based on an original dataset which includes a clustering variable, creates a
list of \code{nboot} datasets with the cluster variable sampled with
replacement.
}
\examples{

my_df <- data.frame(
  id = sample(1:50, size = 500, replace = TRUE),
  x = rnorm(n = 500),
  y = rbinom(n = 500, size = 1, prob = 0.2)
)

create_bootdata(df = my_df, cluster_var = "id", nboot = 25)

}

