% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{eval_formula_with}
\alias{eval_formula_with}
\title{Wrapper to allow formula object to be passed to with([mids object], vglm(...))}
\usage{
eval_formula_with(data, formula, fitter = try_vglm, ...)
}
\arguments{
\item{data}{Object on which you can fit a model of type \code{fitter}. Most
cases, a \code{data.frame} or \code{mids} object.}

\item{formula}{\code{formula}; see \code{\link[stats]{formula}} for more
details.}

\item{fitter}{Modeling function. In this package, defaults to
\code{\link{try.vglm}}.}
}
\description{
Passing a formula object to a with(mids object, ...) call breaks.
eval(substitute()) makes it work.
}

