% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{extract_coefs_s4}
\alias{extract_coefs_s4}
\title{Extract coefficients from an S4 model object that may be of class try-error}
\usage{
extract_coefs_s4(x, ncoefs)
}
\arguments{
\item{x}{either S4 model object or object of type try-error}

\item{ncoefs}{numeric; number of missing values to return if object is try-error}
}
\description{
Extract coefficients from an S4 model object that may be of class try-error
}

