% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_extract_multinom.R
\name{extract_multinom_info.default}
\alias{extract_multinom_info.default}
\title{describeIn extract_multinom_info Method for models not fit with
  \code{\link[mice]{mids}} objects}
\usage{
\method{extract_multinom_info}{default}(modobj, coef_only = TRUE)
}
\description{
describeIn extract_multinom_info Method for models not fit with
  \code{\link[mice]{mids}} objects
}

