% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_extract_multinom.R
\name{extract_multinom_info.mira}
\alias{extract_multinom_info.mira}
\title{describeIn extract_multinom_info Method for models fit with
  \code{\link[mice]{mids}} objects}
\usage{
\method{extract_multinom_info}{mira}(modobj, coef_only = TRUE,
  coef_matrix = FALSE, fail_pct = 0.8)
}
\description{
describeIn extract_multinom_info Method for models fit with
  \code{\link[mice]{mids}} objects
}

