% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_extract_multinom.R
\name{extract_multinom_info}
\alias{extract_multinom_info}
\title{Extract information from a single try_vglm fit}
\usage{
extract_multinom_info(modobj, coef_only = TRUE, ...)
}
\arguments{
\item{modobj}{Object created using \code{\link{try_vglm}}, of class
\code{try-error}, \code{\link[VGAM]{vglm}}, or \code{\link[mice]{mids}}.}

\item{coef_only}{logical; whether to return only model coefficients
(recommended to save memory; in the context of bootstrapping, coefficients
are often the only required information) or entire model object. Defaults
to TRUE.}

\item{coef_matrix}{logical; whether to return a matrix of all coefficient
estimates from each imputation in a \code{mids} object, in addition to the
averages for each coefficient. Defaults to FALSE.}

\item{fail_pct}{numeric, used if \code{df} is a \code{mids} object;
proportion of imputed datasets that can have errors or warnings and still
be able to calculate point estimates. Defaults to 0.8 (80\% of imputations
must fit with no errors or warnings).}
}
\value{
List containing the following elements:
\itemize{
  \item \code{fitsucc}: logical indicating whether the model was successfully fit,
    with no warnings/errors
  \item\code{coefs}: numeric vector of model coefficients (NULL if
    \code{fitsucc} = FALSE)
   \item\code{msgs}: character vector of warnings/error messages (NULL if
    \code{fitsucc} = TRUE)
   \item\code{modobj}: full model object of class \code{VGAM::vglm()}. Only
    returned if requested (\code{coef_only} = FALSE).
}

If \code{df} is a \code{mice::mids} object, the following elements are also
  returned:
\itemize{
  \item \code{nfailsucc}: numeric vector; number of successful and failed model
  fits among imputations
  \item \code{impcoefs} (if \code{coef_matrix = TRUE}): MxP matrix of
    coefficients from each individual imputation
}
}
\description{
This function extracts summary information from a model fit using
\code{\link{try_vglm}} and returns it in a list for easy access.
}
\examples{

my_df <- data.frame(
  id = sample(1:50, size = 500, replace = TRUE),
  x1 = sample(c(NA, 1:100), size = 500, replace = TRUE),
  x2 = sample(c(NA, 1:100), size = 500, replace = TRUE),
  y = sample(c(NA, LETTERS[1:3]), size = 500, replace = TRUE)
)

## Basic usage
my_mod <- fit_multinom(df = my_df, formula = y ~ x1 + x2, ref_level = "A")
extract_multinom_info(my_mod)

## Handle missingness with multiple imputation using mice
## Also try a more complicated formula that might fail to converge
my_df_mice <- mice(my_df)
my_mod_mice <- fit_multinom(
  df = my_df_mice,
  formula = y ~ rcs(x1, 5) * rcs(x2, 5),
  ref_level = "A"
)

extract_multinom_info(my_mod_mice, coef_matrix = TRUE)

}
\seealso{
\code{\link[VGAM]{vglm}}; \code{\link[VGAM]{multinomial}} for model
  fitting. \code{\link[mice]{mice}} for imputation.
}

