% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_extract_multinom.R
\name{fit_extract_multinom}
\alias{fit_extract_multinom}
\title{Fit multinomial model and extract summary info on a single data object}
\usage{
fit_extract_multinom(formula, df, ref_level, ...)
}
\value{
List containing the following elements:
\itemize{
  \item \code{fitsucc}: logical indicating whether the model was successfully fit,
    with no warnings/errors
  \item\code{coefs}: numeric vector of model coefficients (NULL if
    \code{fitsucc} = FALSE)
   \item\code{msgs}: character vector of warnings/error messages (NULL if
    \code{fitsucc} = TRUE)
   \item\code{modobj}: full model object of class \code{VGAM::vglm()}. Only
    returned if requested (\code{coef_only} = FALSE).
}

If \code{df} is a \code{mice::mids} object, the following elements are also
  returned:
\itemize{
  \item \code{nfailsucc}: numeric vector; number of successful and failed model
  fits among imputations
  \item \code{impcoefs} (if \code{coef_matrix = TRUE}): MxP matrix of
    coefficients from each individual imputation
}
}
\description{
This is a wrapper for \code{fit_multinom()} + \code{extract_multinom_info};
it fits a model and extracts summary information (indicator for successful
fit, vector of coefficients, and character vector of errors/warnings) given
a formula, a single data object (\code{data.frame} or \code{mice::mids}
object), and a reference level for the outcome.
}
\examples{

my_df <- data.frame(
  id = sample(1:50, size = 500, replace = TRUE),
  x1 = sample(c(NA, 1:100), size = 500, replace = TRUE),
  x2 = sample(c(NA, 1:100), size = 500, replace = TRUE),
  y = sample(c(NA, LETTERS[1:3]), size = 500, replace = TRUE)
)

## Basic usage
my_mod_info <- fit_extract_multinom(
  formula = y ~ x1 + x2,
  df = my_df,
  ref_level = "A"
)

## Handle missingness with multiple imputation using mice
## Also try a more complicated formula that might fail to converge
my_df_mice <- mice(my_df[1:125,])
my_mod_info_mice <- fit_extract_multinom(
  formula = y ~ rcs(x1, 5) * rcs(x2, 5),
  df = my_df_mice,
  ref_level = "A",
  coef_matrix = TRUE
)

}
\seealso{
\code{\link{fit_multinom}}, \code{\link{extract_multinom_info}}.
  \code{\link[VGAM]{vglm}}; \code{\link[VGAM]{multinomial}} for model
  fitting. \code{\link[mice]{mice}} for imputation.
}

