% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_extract_multinom.R
\name{fit_multinom}
\alias{fit_multinom}
\title{Fit multinomial logistic regression model on a single object}
\usage{
fit_multinom(formula, df, ref_level, fitter = try_vglm)
}
\arguments{
\item{formula}{\code{formula}; see \code{\link[stats]{formula}} for more
details.}

\item{df}{\code{data.frame} or \code{mice::mids} object on which to run model.}

\item{ref_level}{numeric (representing factor level) or character; level of
outcome variable to use as reference.}

\item{fitter}{Function to use for model fitting. Currently only option is
\code{\link{try_vglm}}.}
}
\description{
This function runs a multinomial logistic regression model on a data.frame or
`mice::mids` object, currently using \code{\link{try_vglm}} (which runs
\code{VGAM::vglm(..., family = multinomial(...))}, but if the model fails to
converge - ie, warning or error - the warning/error message is returned
instead).
}
\details{
This wrapper function is necessary to perform this operation on
\code{\link[mice]{mids}} objects, which require a \code{with(..., )} syntax.
}
\examples{

my_df <- data.frame(
  id = sample(1:50, size = 500, replace = TRUE),
  x1 = sample(c(NA, 1:100), size = 500, replace = TRUE),
  x2 = sample(c(NA, 1:100), size = 500, replace = TRUE),
  y = sample(c(NA, LETTERS[1:3]), size = 500, replace = TRUE)
)

## Works whether df is a data.frame or mice() object
fit_multinom(y ~ x1 + x2, my_df, ref_level = "A")
fit_multinom(y ~ x1 + x2, mice::mice(my_df), ref_level = "A")
}
\seealso{
\code{\link[VGAM]{vglm}}; \code{\link[VGAM]{multinomial}} for model
  fitting. \code{\link[mice]{mice}} for imputation.
}

