% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{reformat_vnames}
\alias{reformat_vnames}
\title{Reformat a vector of strings (eg, coefficient names from VGAM::vglm()) to
conform with tibble naming conventions}
\usage{
reformat_vnames(vnames)
}
\arguments{
\item{vnames}{character vector}
}
\description{
Conventions:
}
\details{
- Coefficients for each outcome level designated by \code{_x}, where \code{x}
indicates the outcome level (eg, \code{x1_1, x1_2})
- Intercept terms designated by \code{_by_} (replaces \code{":"})
- Nonlinear terms designated by \code{"."} (eg, \code{x1_1, x1._1})
}

