% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multinom.R
\name{run_multinom.default}
\alias{run_multinom.default}
\title{describeIn run_multinom Method for data.frames}
\usage{
\method{run_multinom}{default}(df, formula, ref_level, coef_only = TRUE)
}
\description{
describeIn run_multinom Method for data.frames
}

