% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multinom.R
\name{run_multinom.mids}
\alias{run_multinom.mids}
\title{describeIn run_multinom Method for \code{mice::mids} objects}
\usage{
\method{run_multinom}{mids}(df, formula, ref_level, coef_only = TRUE,
  coef_matrix = FALSE, fail_pct = 0.8)
}
\description{
describeIn run_multinom Method for \code{mice::mids} objects
}

