% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multinom.R
\name{run_multinom}
\alias{run_multinom}
\title{Run multinomial logistic regression model on a single object}
\usage{
run_multinom(df, formula, ref_level, coef_only, ...)
}
\arguments{
\item{df}{\code{data.frame} or \code{mice::mids} object on which to run model.}

\item{formula}{character string that can be coerced to \code{formula}; see
\code{\link[stats]{formula}} for more details.}

\item{ref_level}{numeric (representing factor level) or character; level of
outcome variable to use as reference.}

\item{coef_only}{logical; whether to return entire model object or only model
coefficients (recommended to save space; in the context of bootstrapping,
coefficients are often the only required information). Defaults to TRUE.}

\item{coef_matrix}{logical; whether to return a matrix of all coefficient
estimates from each imputation in a \code{mids} object, in addition to the
averages for each coefficient. Defaults to FALSE.}

\item{fail_pct}{numeric, used if \code{df} is a \code{mids} object;
proportion of imputed datasets that can have errors or warnings and still
be able to calculate point estimates. Defaults to 0.8 (80\% of imputations
must fit with no errors or warnings).}
}
\value{
List containing the following elements:
\itemize{
  \item \code{fitsucc}: logical indicating whether the model was successfully fit,
    with no warnings/errors
  \item\code{coefs}: numeric vector of model coefficients (NULL if
    \code{fitsucc} = FALSE)
   \item\code{msgs}: character vector of warnings/error messages (NULL if
    \code{fitsucc} = TRUE)
   \item\code{modobj}: full model object of class \code{VGAM::vglm()}. Only
    returned if requested (\code{coef_only} = FALSE).
}

If \code{df} is a \code{mice::mids} object, the following elements are also
  returned:
\itemize{
  \item \code{nfailsucc}: numeric vector; number of successful and failed model
  fits among imputations
  \item \code{impcoefs}: MxP matrix of coefficients from each individual imputation
}
}
\description{
This function runs a multinomial logistic regression model on a data.frame or
`mice::mids` object, currently using
\code{VGAM::vglm(..., family = multinomial(...))}. The user can specify
whether to save the entire \code{vglm()} object or only the coefficients. If
the model fails to converge (warning or error), the warning/error message is
returned instead.
}
\examples{

my_df <- data.frame(
  id = sample(1:50, size = 500, replace = TRUE),
  x1 = sample(c(NA, 1:100), size = 500, replace = TRUE),
  x2 = sample(c(NA, 1:100), size = 500, replace = TRUE),
  y = sample(c(NA, LETTERS[1:3]), size = 500, replace = TRUE)
)

## Basic usage
run_multinom(df = my_df, formula = "y ~ x1 + x2", ref_level = "A")

## Handle missingness with multiple imputation using mice
## Also try a more complicated formula that might fail to converge
my_df_mice <- mice(my_df)
run_multinom(
  df = my_df_mice,
  formula = "y ~ rcs(x1, 5) * rcs(x2, 5)",
  ref_level = "A",
  coef_matrix = TRUE
)

}
\seealso{
\link[VGAM]{vglm}; \link[VGAM]{multinomial}; \link[mice]{mice}
}

