% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{sample_clusters}
\alias{sample_clusters}
\title{Return a sample of a vector with replacement, given a seed}
\usage{
sample_clusters(use_seed, use_clusters)
}
\arguments{
\item{use_seed}{numeric; seed to set}

\item{use_clusters}{vector to sample from}
}
\description{
Return a sample of a vector with replacement, given a seed
}

