% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_multinom_list.R
\name{summarize_multinom_list}
\alias{summarise_multinom_list}
\alias{summarize_multinom_list}
\title{Fit multinomial model and extract summary info on a list of data objects}
\usage{
summarize_multinom_list(formula, df_list, ref_level, orgdf = NULL,
  orginfo = c("coefs", "nothing", "modobj"), nsuccfits = NULL, ...)

summarise_multinom_list(formula, df_list, ref_level, orgdf = NULL,
  orginfo = c("coefs", "nothing", "modobj"), nsuccfits = NULL, ...)
}
\arguments{
\item{formula}{\code{formula}; see \code{\link[stats]{formula}} for more
details.}

\item{df_list}{\code{list} of either \code{data.frame} or \code{mice::mids}
objects on which to run model.}

\item{ref_level}{numeric (representing factor level) or character; level of
outcome variable to use as reference.}

\item{orgdf}{*(optional)* \code{data.frame} or \code{mice::mids} object to
"test" the model. If the model does not converge successfully with this
data, the function will stop. (Example: If \code{df_list} is a list of
bootstrapped data frames, \code{orgdf} would likely be the data frame from
which the bootstraps come.)}

\item{orginfo}{*(optional)*, if \code{orgdf} supplied; one of \code{coefs}
(default), \code{nothing}, or \code{modobj}. If \code{coefs}, returns only
a named vector of coefficients. If \code{modobj}, returns entire model fit
on \code{orgdf}. If \code{nothing}, returns nothing; only uses \code{orgdf}
to test whether model runs successfully.}

\item{nsuccfits}{*(optional)* \code{numeric}; number of successful fits
requested. If specified, and more than \code{nsuccfits} successful model
fits result from \code{df_list}, only the first \code{nsuccfits} will be
returned in the final function information. If specified and fewer than
\code{nsuccfits} are available, all information will be returned and a
warning message will display.}

\item{...}{Additional arguments to pass to \code{fit_extract_multinom()}.}
}
\value{
List containing the following elements:
\itemize{
  \item \code{allresults}: \code{\link{[tibble]tibble}} containing one row
    per element of \code{df_list}, with the following information for each
    included as variables or [list columns](https://www.rstudio.com/resources/videos/using-list-cols-in-your-dataframe/):
    \itemize{
      \item \code{fitsucc}: logical; indicates whether model was successfully
        fit to that data object
      \item \code{coefs}: tibble with one row, and one column per coefficient
      \item \code{impcoefs}, if requested: tibble with coefficient estimates
        from each imputation
      \item \code{msgs}: list of character strings; errors/warning messages
        from unsuccessful fits
      \item \code{modobj}, if requested: full model object
    }
  \item \code{fitsucc}: numeric vector including number of model fit Successes,
    Failures, and (if applicable) Imputation Successes/Failures
  \item \code{coefs}: tibble with one row per element of \code{df_list} and one
    column per model coefficient (no rows for unsuccessful model fits)
  \item \code{impcoefs}, if requested: tibble with one row per *imputation* of
    \code{df_list} with a successful model fit and one column per model
    coefficient (no rows for unsuccessful model fits)
  \item \code{orgcoefs}, if \code{orgdf} is supplied and
    \code{orginfo = "coefs"}: named vector of coefficients from model fit on
    starting data frame (eg, original data frame from which \code{df_list} is
    bootstrapped)
  \item \code{orgmod}, if \code{orgdf} is supplied and
    \code{orginfo = "modobj"}: full model fit on starting data frame (eg,
    original data frame from which \code{df_list} is bootstrapped)
}
}
\description{
This function calls \code{fit_extract_multinom()} (which calls
\code{fit_multinom()} + \code{extract_multinom_info}) on a list of data
objects (\code{data.frame} or \code{\link[mice]{mice}}), then summarizes the
results of each.
}
\examples{

my_df <- data.frame(
  id = sample(1:50, size = 500, replace = TRUE),
  x1 = sample(c(NA, 1:100), size = 500, replace = TRUE),
  x2 = sample(c(NA, 1:100), size = 500, replace = TRUE),
  y = sample(c(NA, LETTERS[1:3]), size = 500, replace = TRUE)
)

## Basic usage
my_dflist <- create_bootdata(my_df, cluster_var = "id", nboot = 10)

my_mod_summary <- summarize_multinom_list(
  formula = y ~ x1 + x2,
  df_list = my_dflist,
  ref_level = "A",
  orgdf = my_df
)

## Supply original data frame as testdf; see what happens when model is too
## complex
my_df_small <- my_df[1:100,]
my_dflist_small <- create_bootdata(my_df_small, cluster_var = "id", nboot = 10)

my_mod_summary <- summarize_multinom_list(
  formula = y ~ rms::rcs(x1, 4) * rms::rcs(x2, 4),
  df_list = my_dflist_small,
  ref_level = "A",
  orgdf = my_df_small
)

## Handle missingness with multiple imputation using mice
## Also try a more complicated formula that might fail to converge
my_dflist_mice <- create_bootdata(
  my_df[1:100,], cluster_var = "id", nboot = 10, impute = TRUE
)

my_mod_info_mice <- summarize_multinom_list(
  formula = y ~ rcs(x1, 4) * rcs(x2, 4),
  df_list = my_dflist_mice,
  ref_level = "A",
  coef_matrix = TRUE
)

}
\seealso{
\code{\link{fit_multinom}}, \code{\link{extract_multinom_info}},
  \code{\link{fit_extract_multinom}}. \code{\link[VGAM]{vglm}};
  \code{\link[VGAM]{multinomial}} for model fitting. \code{\link[mice]{mice}}
  for imputation.
}

