% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{try.vglm}
\alias{try.vglm}
\title{Silently catch errors for \code{vglm()} calls}
\usage{
try.vglm(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{VGAM::vglm()}.}
}
\description{
Silently catch errors for \code{vglm()} calls
}

