% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_vglm.R
\name{try_vglm}
\alias{try_vglm}
\title{Silently catch errors for \code{vglm()} calls}
\usage{
try_vglm(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{VGAM::vglm()}.}
}
\description{
Silently catch errors for \code{vglm()} calls
}

